backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.9.0
Werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flasgger==0.9.7.1
